% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinompars.R
\name{BetaBinompars}
\alias{BetaBinompars}
\title{Estimate parameters from Beta Binomial model}
\usage{
BetaBinompars(VAF)
}
\arguments{
\item{data}{VAF data: contains Locus,chrom, vc, dp, smapleID}
}
\value{
rslt	Beta Binomail parameter estimates for each locus each group
}
\description{
Estimate parameters from Beta Binomial model
}
\examples{
data(VAF)
temp <- BetaBinompars(VAF)
summary(temp)

}
\references{
{

  \insertRef{aod}{easyVAF}

}
}
