% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{betapar}
\alias{betapar}
\title{Title helper function: betapar}
\usage{
betapar(mu, var)
}
\arguments{
\item{mu}{mean of VAF}

\item{var}{variance of VAF}
}
\value{

}
\description{
Title helper function: betapar
}
\keyword{internal}
