% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{power.vaf}
\alias{power.vaf}
\title{Title}
\usage{
power.vaf(mus, vars, n, N, nsim, alpha.level)
}
\arguments{
\item{mus}{means of groups's VAF}

\item{vars}{variance of VAF for different groups}

\item{n}{read depth}

\item{N}{sample size}

\item{nsim}{number of simulations}

\item{alpha.level}{}
}
\value{
simulation result lists including "prob.reject.fisher", "prob.reject.pearson",  "prob.reject.betabinom", and so on
}
\description{
Title
}
\examples{
power.vaf(c(0.2, 0.2), c(0.01, 0.01), 50, 3, 100, 0.05)

}
