% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAFmain.R
\name{QCchecking}
\alias{QCchecking}
\title{Quality checking for biological variability among samples}
\usage{
QCchecking(data, method = "lm")
}
\arguments{
\item{data}{VAF data: contains locus,chrom, vc, dp, smapleID}

\item{method}{choose from "lm" or "lmer"}
}
\value{
rslt QC result table
}
\description{
Quality checking for biological variability among samples
}
\examples{
data(dat)
rslt <- QCchecking(data=VAF, method="lm")

}
\references{
{
  \insertRef{ggplot2}{easyVAF}

  \insertRef{lme4}{easyVAF}

}
}
