% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simdat}
\alias{simdat}
\title{Title helper function of simdat}
\usage{
simdat(mus, vars, n, N)
}
\arguments{
\item{mus}{means of VAF}

\item{vars}{variances of VAF}

\item{n}{readdepth}

\item{N}{sample size}
}
\value{

}
\description{
Title helper function of simdat
}
\keyword{internal}
