% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taronetest.R
\name{Tarone.test}
\alias{Tarone.test}
\title{Tarone test: overdispersion tests}
\usage{
Tarone.test(N, M)
}
\arguments{
\item{N}{Number of trials}

\item{M}{Observed count value}
}
\value{
Tarone test result
}
\description{
Tarone tes is testing for overdispersion in Poisson and Binomial Regression Models.
}
\examples{
N <- c(30, 32, 40, 28, 29, 35, 30, 34, 31, 39)
M <- c( 9, 10, 22, 15,  8, 19, 16, 19, 15, 10)
Tarone.test(N, M)
}
\references{
\url{https://stats.stackexchange.com/a/410376/6378} and
R. E. TARONE, Testing the goodness of fit of the binomial distribution, Biometrika, Volume 66, Issue 3, December 1979, Pages 585–590, \url{https://doi.org/10.1093/biomet/66.3.585}
}
\author{
\href{https://stats.stackexchange.com/users/173082/reinstate-monica}{Ben O'Neill}
}
